Refactor ingredient logic component into its own package
