+++ src/main/java/fridgy/logic/commands/ingredient/ListCommand.java
@@ -1,7 +1,9 @@
-package fridgy.logic.commands;
+package fridgy.logic.commands.ingredient;
 
 import static java.util.Objects.requireNonNull;
 
+import fridgy.logic.commands.Command;
+import fridgy.logic.commands.CommandResult;
 import fridgy.model.IngredientModel;
 import fridgy.model.Model;
 import fridgy.model.ingredient.IngredientDefaultComparator;
