+++ src/main/java/fridgy/logic/parser/ingredient/AddCommandParser.java
@@ -1,4 +1,4 @@
-package fridgy.logic.parser;
+package fridgy.logic.parser.ingredient;
 
 import static fridgy.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static fridgy.logic.parser.CliSyntax.PREFIX_DESCRIPTION;
@@ -14,7 +14,12 @@ import java.util.regex.Pattern;
 import java.util.stream.Stream;
 
 import fridgy.commons.core.Messages;
-import fridgy.logic.commands.AddCommand;
+import fridgy.logic.commands.ingredient.AddCommand;
+import fridgy.logic.parser.ArgumentMultimap;
+import fridgy.logic.parser.ArgumentTokenizer;
+import fridgy.logic.parser.Parser;
+import fridgy.logic.parser.ParserUtil;
+import fridgy.logic.parser.Prefix;
 import fridgy.logic.parser.exceptions.ParseException;
 import fridgy.model.ingredient.Description;
 import fridgy.model.ingredient.ExpiryDate;
