+++ src/test/java/fridgy/logic/commands/AddCommandIntegrationTest.java
@@ -3,6 +3,7 @@ package fridgy.logic.commands;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import fridgy.logic.commands.ingredient.AddCommand;
 import fridgy.model.Model;
 import fridgy.model.ModelManager;
 import fridgy.model.RecipeBook;
