+++ src/test/java/fridgy/logic/parser/EditCommandParserTest.java
@@ -5,7 +5,8 @@ import org.junit.jupiter.api.Test;
 import fridgy.commons.core.Messages;
 import fridgy.commons.core.index.Index;
 import fridgy.logic.commands.CommandTestUtil;
-import fridgy.logic.commands.EditCommand;
+import fridgy.logic.commands.ingredient.EditCommand;
+import fridgy.logic.parser.ingredient.EditCommandParser;
 import fridgy.model.ingredient.Description;
 import fridgy.model.ingredient.Name;
 import fridgy.model.ingredient.Quantity;
