+++ src/test/java/fridgy/logic/parser/InventoryParserTest.java
@@ -11,16 +11,18 @@ import java.util.stream.Collectors;
 import org.junit.jupiter.api.Test;
 
 import fridgy.commons.core.Messages;
-import fridgy.logic.commands.AddCommand;
 import fridgy.logic.commands.ClearCommand;
 import fridgy.logic.commands.CommandTestUtil;
-import fridgy.logic.commands.DeleteCommand;
-import fridgy.logic.commands.EditCommand;
-import fridgy.logic.commands.EditCommand.EditIngredientDescriptor;
-import fridgy.logic.commands.FindCommand;
 import fridgy.logic.commands.HelpCommand;
-import fridgy.logic.commands.ListCommand;
+import fridgy.logic.commands.ingredient.AddCommand;
+import fridgy.logic.commands.ingredient.DeleteCommand;
+import fridgy.logic.commands.ingredient.EditCommand;
+import fridgy.logic.commands.ingredient.EditCommand.EditIngredientDescriptor;
+import fridgy.logic.commands.ingredient.FindCommand;
+import fridgy.logic.commands.ingredient.ListCommand;
 import fridgy.logic.parser.exceptions.ParseException;
+import fridgy.logic.parser.ingredient.AddCommandParser;
+import fridgy.logic.parser.ingredient.InventoryParser;
 import fridgy.model.ingredient.ExpiryDate;
 import fridgy.model.ingredient.Ingredient;
 import fridgy.model.ingredient.Name;
