+++ src/test/java/fridgy/logic/parser/ViewCommandParserTest.java
@@ -3,7 +3,8 @@ package fridgy.logic.parser;
 import org.junit.jupiter.api.Test;
 
 import fridgy.commons.core.Messages;
-import fridgy.logic.commands.ViewCommand;
+import fridgy.logic.commands.ingredient.ViewCommand;
+import fridgy.logic.parser.ingredient.ViewCommandParser;
 
 public class ViewCommandParserTest {
